tests/wpt/tests/navigation-api/navigation-history-entry/no-referrer-dynamic-url-censored.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async (t) => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(r => window.onload = () => t.step_timeout(r, 0));

  // The entry for the first document has a visible url.
  i.contentWindow.navigation.navigate("/common/blank.html?2");
  await new Promise(r => i.onload = () => t.step_timeout(r, 0));
  assert_not_equals(i.contentWindow.navigation.entries()[0].url, null);
  assert_not_equals(i.contentWindow.navigation.activation.from.url, null);

  // Apply no-referrer, the url should now be censored when no longer on that document.
  i.contentWindow.navigation.back();
  await new Promise(r => i.onload = () => t.step_timeout(r, 0));
  i.contentDocument.head.innerHTML = `<meta name="referrer" content="no-referrer">`;
  assert_not_equals(i.contentWindow.navigation.entries()[0].url, null);
  i.contentWindow.navigation.forward();
  await new Promise(r => i.onload = () => t.step_timeout(r, 0));
  assert_equals(i.contentWindow.navigation.entries()[0].url, null);
  assert_equals(i.contentWindow.navigation.activation.from.url, null);

  // Overwrite the referrer policy, the url should be visible again.
  i.contentWindow.navigation.back();
  await new Promise(r => i.onload = () => t.step_timeout(r, 0));
  i.contentDocument.head.innerHTML = `<meta name="referrer" content="same-origin">`;
  i.contentWindow.navigation.forward();
  await new Promise(r => i.onload = () => t.step_timeout(r, 0));
  assert_not_equals(i.contentWindow.navigation.entries()[0].url, null);
  assert_not_equals(i.contentWindow.navigation.activation.from.url, null);
}, "The url of a document is censored by a no-referrer policy dynamically");
</script>
