LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/no-referrer-dynamic-url-censored-expected.txt



FAIL The url of a document is censored by a no-referrer policy dynamically assert_equals: expected (object) null but got (string) "http://localhost:8800/common/blank.html"
